feat: add modal
